requirements.txt
SQLAlchemy==2.0.25
requests==2.31.0
python-dotenv==1.0.0
urllib3==2.0.7
PyJWT==2.8.0
cryptography==41.0.5
pydantic==2.5.2